.history/unit_tests/requirements_20251211033918.txt
# Core testing framework
pytest>=7.4.0
pytest-mock>=3.11.1

# Parallel execution
pytest-xdist>=3.3.1
pytest-parallel>=0.1.1

# Reporting
pytest-html>=3.2.0
pytest-cov>=4.1.0
allure-pytest>=2.13.2

# Code quality
pylint>=2.17.5
flake8>=6.0.0

# BDD framework
behave>=1.2.6

# Additional utilities
pytest-timeout>=2.1.0
pytest-json-report>=1.5.0
